requirements.txt
ccxt>=4.3.0
numpy<2.0,>=1.23.5
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.4.2
requests>=2.32.0
python-dotenv>=1.0.1
ta>=0.11
